Use only global ip addr for ttyd
After installing lxqt, for unknown reasons, avahi reports ip addr(ipv4,
ipv6) separately(before that, it reports them together). If it reports
ipv6 addr, it is problematic because reported ipv6 addr is link-local
addr which is not availble in webview.(it doesn't support the schema for
scope id). So let it use 'global' ip addr only(which is ipv4 addr for
our cases)

Bug: 438045731
Test: install lxqt, and restart terminal
Flag: EXEMPT a simple bug fix
Change-Id: I0f11ce69be0bd9cbb2a8626edf61ddcca50757a5