testMinimumRequiredRAM: measure RSS instead of requested RAM
The requested RAM size is misleading because AVF automatically adds
protected VM overheads to the total, including the pvmfw region and
swiotlb.

Before:

    [3/6] com.android.microdroid.benchmark.MicrodroidBenchmarks#testMinimumRequiredRAM[protectedVm=true,os=microdroid]: PASSED (22.647s)
            avf_perf[hwrev_a0]/microdroid/minimum_required_memory: 32
    [4/6] com.android.microdroid.benchmark.MicrodroidBenchmarks#testMinimumRequiredRAM[protectedVm=false,os=microdroid]: PASSED (44.754s)
            avf_perf[hwrev_a0]/microdroid/minimum_required_memory: 35

After:

    [3/6] com.android.microdroid.benchmark.MicrodroidBenchmarks#testMinimumRequiredRAM[protectedVm=true,os=microdroid]: PASSED (25.834s)
            avf_perf[hwrev_a0]/microdroid/minimum_required_memory: 42
    [4/6] com.android.microdroid.benchmark.MicrodroidBenchmarks#testMinimumRequiredRAM[protectedVm=false,os=microdroid]: PASSED (55.622s)
            avf_perf[hwrev_a0]/microdroid/minimum_required_memory: 35

Switching to "assets/vm_config_io.json" for the guest and connecting to
the benchmark vsock service adds a bit of bloat that isn't needed, but
it is convenient for sharing code with the other benchmarks. Since the
non-protected case didn't increase, the effect seems to be small enough
to ignore.

Test: atest "MicrodroidBenchmarkApp:com.android.microdroid.benchmark.MicrodroidBenchmarks#testMinimumRequiredRAM"
Flag: EXEMPT .
Change-Id: Ifcc162308816489c00215e54d3054d23e9935da9