VirtualMachine.java: don't orphan VM when "onDied" callback is delayed
It is possible to make a sequence of API calls like the following and
have the `onDied` binder call from virtmgr arrive *after* the second
`vm.run()` call.

    VirtualMachine vm = ...;
    vm.run();    // start VM
    vm.close();  // stop VM
    vm.run();    // start a new VM

In that case, the `VirtualMachine` object is managing a live VM and
`handleStopped` must not just drop the reference to it, otherwise we'll
orphan the VM and clients can end up with two (or more) copies of the
same VM configuration running concurrently, likely causing corruption in
the VM's writable disk images.

The fix is to always pass a target CID to `handleStopped`. If the CID
doesn't match the most recently launched VM, then the notication is
stale and we can ignore it.

Bug: 444733142
Test: see bug
Flag: NONE .
Change-Id: I2234d1c226737ad0fdad95306b3ab8986b09dd17