VirtualMachine.java: rework callback handling
A VirtualMachine instance can be used to spawn multiple VMs, but only
one at a time. If a VM is stopped and a new one started, sometimes the
callbacks for the old VM get process after the new VM has started and,
if `setCallback` was used to register new callbacks for the new VM, we
incorrectly route the old VM's events to the new callbacks.

To make the whole situation more robust, track the callbacks directly in
`CallbackTranslator`. `CallbackTranslator` is the actual binder object
registered to a particular VM and so will only receive events for that
specific VM. Calls to `setCallback` will only update the
`CallbackTranslator` of the most recently spawned VM so that any late
arriving events for an older VM will continue to be routed to the
correct callback.

Bug: 444733142
Test: its complicated, see the bug
Flag: NONE .
Change-Id: Ibc55b912ab2ee81c7b121e5caa3258b62e7052ab