pvmfw: Allow an alternate digest for kernel verification.
AL sllow booting from a USB drive when in dev mode. When this happens we
need to boot an alternate TEE image that will not alter the tamper
evident storage on the internal storage. The current change allows
either kernel to boot, however once b/445576875 adds a sysprop that
indicates internal versus external boot this will be updated to check
the proper hash based on the boot mode.

Flag: build.RELEASE_TRUSTY_GATEKEEPER_ENABLED
Bug: 441097688
Test: boot from USB, android-desktop-install, boot from internal storage

Change-Id: Ica321799e850a6dec35fc2abf8b19b2e93d31eb3