Kill virtualizationservice upon SIGTERM
Previously, it waited (up to 3 seconds) for init to forcibly kill it via
SIGKILL. That unnecessarily delayed the shutdown process.

Flag: EXEMPT BUGFIX
Bug: 450174930
Test: packages/modules/Virtualization/android/vm/vm_shell.sh
start-microdroid --auto-connect
And after the VM is up, reboot Android. And observe logcat.
virtualizationservice is killed soon after it receives SIGTERM, and
it is not listed in the "still aliving" processes by init.

Change-Id: Ia3e1a5ca1964175f7bd4a39b133755314346b4e5