payload: support block-file APEX
Creating a payload fails when one of its components is a block file
because it fails to get the size. crosvm fails to open the payload
(composite disk) because of two partitions start at the same offset.

This happens with "mount_before_data" feature, which uses block devices
for updated APEXes.

In this change, the size of a block file is retrieved via BLKGETSIZE64
ioctl.

Bug: 455685757
Test: odsign_e2e_test
Flag: com.android.apex.flags.mount_before_data
Change-Id: I8e69e883a06ce46098e0e185ef8939f89b47bc21