Add arm64.nompam when launching ferrochrome VM
KVM, until 6.12, did not support MPAM while it was exposed to the guest.
To work around this bug, disable MPAM on the guest side.

When Ferrochrome kernel switches to GKI, we will no longer need this.

Also note that this new argument is added also to x86_64 kernel, but it
is ignored there.

Bug: 456105963
Flag: EXEMPT BUGFIX
Test: N/A
Change-Id: I3d8e3f6e07b0b57ad50df7e2b15d3ec76f69871a