Add preinstall allowlisting for apps in the virt APEX
Terminal app: installed for FULL users and managed profiles. Note
however that this doesn't mean that the app is shown in the launcher by
default for all users. The user has to explicitly enable the "Linux
terminal" feature in the developer options. Then the main activity of
the terminal app is enabled for that user.

EmptyPayload: installed only for the SYSTEM user. This apk is basically
to run the testing-purpose Microdroid VM. Since we don't intend to run
Microdroid tests on any other users, this can be installed only to the
SYSTEM user.

ComposPayload: installed only for the SYSTEM user. This app is to run
the isolated compilation which compiles updated framework jar libs into
native code. Since the compilation is a system-internal task and is done
on behalf of the "system" user, we don't have to install it to other
types of users.

Bug: 451960740
Bug: 451959811
Bug: 451961058
Bug: 457388297
Test: check that three apps are installed on the default user
Test: add a user, and check that terminal app is installed but others
are not.
Flag: EXEMPT BUGFIX

Change-Id: I9288eba3e29aa5267c895abcbbed18e6270fc8c2