liblibfdt: Introduce find_or_add_node_mut()
Introduce a helper function to locate a node from a given path, or
create it (and any needed parent) if it doesn't already exist.

Use the function in pvmfw, in prevision of the refactor of the support
for the VM ref DT, to support /firmware/android.

Use the helper in virtmgr to make it more robust (some corner cases
might have previously failed with the wrong set of inputs).

Bug: 454303624
Test: m pvmfw_bin service_vm_bin virtualizationmanager_device_test
Test: m libpvmfw.device_assignment.test vmbase_example_bios_bin
Flag: EXEMPT PURE_REFACTOR
Change-Id: I3a15e5624eb2b8097611d1a7df7f14046317273f