Use empty udroid payload for testNoLongHypSections and testPsciMemProtect
Previously, we used compos as a payload for hypervisor testing. Even
though it is an interesting real use-case it however has downsides.
First it is quite a long process and it will generate MiBs of events.
Second, it is only possible to run a single instance of that VM.

Replace it with a simple boot/shutdown of a Microdroid test instead.

Bug: 442685333
Test: atest AVFHostTestCase#testNoLongHypSections
Test: atest AVFHostTestCase#testPsciMemProtect
Flag: EXEMPT TEST_ONLY
Change-Id: I849124ddd410f736798dcac207d51ccf63009ee5