pvmfw: fallback to force_new_instance if instance.img is missing
Currently the `perform_rollback_protection` function expects that pVM is
either:
* a well-known pVM (e.g. rkpvm);
* knows how to handle the deferred rollback protection;
* or has instance.img virtio-blk device.

If none of the assumptions is true, then pvmfw rejects the pVM boot.
This is a bit strict for baremetal pVMs, hence this patch.

This patch changes the behaviour from failing the pVM boot to simply
fallback to the force_new_instance flow. It is not perfect, but should
be enough to unblock the VTS test migration from service_vm to the
dedicated vmbase_test_vm.

Bug: 457714108
Test: presubmit
Test: m
Flag: EXEMPT UNSUPPORTED_LANGUAGE
Change-Id: Icfa562333dab72eb02bc4570912fe3d8d52b57b1