Set debian network adapters as optioanl
On debian 13, it's observed that the network adapter is initialized as
"eth0" and then renamed to "enp0s13" on the first boot. Starting from
the second boot, systemd-networkd-wait-online incorrectly waits for
"eth0" and causes a delay of 2 minutes.

This can be solved by marking those network adapters as optional. That's
fine in terms of Ferrochrome because connection between the guest VM and
the host Android is important and that has nothing to do with the
internet access itself.

Flag: EXEMPT BUGFIX

Bug: 426073428
Test: build kernel, build trixie, and run terminal app
Change-Id: Ic4698ec64b5a653429a9bd1594383d1513cb9e05