Better error handling
An error is handled in three different ways: show a snackbar leading to
the network settings page, show a snackbar providing an option to retry,
and show a snackbar leading to the bugreporting tool.

Bug: 448356313
Test: delete the VM image. turn the wifi off. start the app. the app
shows a snackbar. turn the wifi on.
Flag: com.android.system.virtualmachine.flags.terminal_newui_jetpack

Change-Id: Iec82e1e67d8de8cffdc69abd28ebe4eb1683615d