Gracefully handle the failure to signal the readiness
virtmgr signals its readiness to the VM client by sending a single
character to the pipe that the client has given. Previously, if it fails
to write to it, it panicked. But this may be too catastrophic, because
the VM client may have been killed by LMK or some other reasons.

Instead of panicking, let's log the incident (with a more informative
message) and exit.

Bug: 451517451
Test: N/A
Flag: EXEMPT BUGFIX
Change-Id: I000af3d4f37f44f494c59b870e46c7a82e258543