Ignore SIGTERM for virtmgr and crosvm
There's a race that can happen during shutdown that is observed in CI
with Cuttlefish. During shutdown, VMs are asked to shut down early
before init starts sending SIGTERM to processes. On device, no issues
have been observed because devices are fast enough. In nested
virtualization environments, that's not the case, and the following can
happen:

- "shutdown.requested" sysprop is read, and virtualizationservice
  requests VMs to stop
- virtmgr sends the stop request
- some time passes before that request is handled due to the slow
  enviornment (in CI, it's 3x nested)
- init kicks off and sends SIGTERM to processes.
- init sends SIGTERM to untracked processes (virtmgr falls here).
- virtmgr does not have a SIGTERM handler and thus is killed.
- VMs are killed as they also receive SIGTERM.

This patch looks to add a SIGTERM handler to allow for VMs to shutdown.
If they are not shutdown directly, init will still SIGKILL them, leading
to the same result. The goal is to provide more time to reduce the
number of kills and avoid having FSCK repairs happen on next boot.

Bug: 465818219
Bug: 440691733
Test: boot CF, shutdown CF, inspect logcat for "Virtmgr stopping"
Flag: EXEMPT bugfix
Change-Id: Ia65b293737fd9b3b1a8d203d466e478a91b918db