old_images_avf_test: don't use service_vm Reverse API
We shouldn't have test-only APIs in production VMs unless it is
absolutely necessary. In case of service_vm there is no need for it,
since we can use vmbase_test_vm for (e.g. VTS) testing.

With this patch the old_images_avf_test becomes duller, since it just
starts a VM and then immediately sends the shutdown request to it ¯\_(ツ)_/¯

Test: atest old_images_avf_test
Flag: EXEMPT TEST_ONLY
Change-Id: I9bc9a730f2f23b9ff8c940fcd4d2e7c4b8a23e64