Forcibly update the VM display when the surface is re-created
When the VM is rendering a static page, the display surface is not
redrawn even when the surface is re-created (due to rotation, etc.) This
is because from the perspective of VM, nothing has changed. But as a
result, the display surface becomes black after re-creation.

This CL fixes that by triggering a redrawning on the VM side. This is
done by sending a mouse jiggle (move pointer 1px, and then come back).

Bug: N/A
Flag: EXEMPT BUGFIX
Test: run gimp, rotate the screen
Change-Id: I470ef28da929b67ebc6e94d193f1da6fc9abde2e